+++ toys/CHANGELOG.md
@@ -6,6 +6,7 @@
 * ADDED: New rake template that supports loading rake tasks as tools.
 * ADDED: Files named ".preload.rb" and files in a ".preload" directory are loaded before tools are defined.
 * ADDED: Directories named ".data" can contain data files accessible from tools.
+* ADDED: Passing "--tools" displays just the list of subtools of a tool
 * IMPROVED: The tool directive can now take an array as the tool name.
 * IMPROVED: The tool directive can now take an `if_defined` argument.
 * FIXED: Template instantiation was failing if the hosting tool was priority-masked.